[IGNORE] Refactor to have a single PrometheusDatasourceSpec
